+++ mex/editor/client/src/app/shared/to-lookup-pipe.spec.ts
@@ -1,0 +1,8 @@
+import { ToLookupPipe } from "./to-lookup-pipe";
+
+describe("ToLookupPipe", () => {
+  it("create an instance", () => {
+    const pipe = new ToLookupPipe();
+    expect(pipe).toBeTruthy();
+  });
+});