+++ CHANGELOG.md
@@ -52,15 +52,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - typedoc to generate typescript documentation
 - docs to typescript components
 - renovate rule for angular
-- fasttrack create resource page with hackathon cleaned-up results
+- fasttrack create resource page
+  - including basic fields only
+- mex-fieldset component to mimic look of kernUx fieldset
 - more linting rules
   - no commented out code
   - no unused imports in component (checked against template)
   - exported members must be documented
+- pipes for typeToIconName, toLabel and toLookup
+- new control fasttrack reference-select that allows to search for existing entites or create them.
 
 ### Changes
 
 - BREAKING: renamed from mex-editor-ng to mex-editor
+- ConceptOptions now use the new ToLookupPipe (which uses the LabelPipe, moved all lookup and label creation to spefic pipes).
+- Grouped everything into fasttrack folder to meat the feature approach. Some components might be moved to shared
 - new template https://github.com/robert-koch-institut/mex-template/releases/tag/1.3.0
 - updated angular to 22.0.4
 - updated template to https://github.com/robert-koch-institut/mex-template/commit/0d8c47